Move project to solution
Moved main code to SpeiderappAPI-folder. This allows adding a test-folder. These are projects under the solution.

See-also: https://docs.microsoft.com/en-us/dotnet/core/porting/project-structure#keep-existing-projects-and-create-a-net-core-project
