+++ SpeiderappAPI.sln
@@ -1,0 +1,16 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SpeiderappAPI", "SpeiderappAPI\SpeiderappAPI.csproj", "{7B80FBC1-DDFD-4A52-84FF-1E978278C760}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{7B80FBC1-DDFD-4A52-84FF-1E978278C760}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7B80FBC1-DDFD-4A52-84FF-1E978278C760}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7B80FBC1-DDFD-4A52-84FF-1E978278C760}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7B80FBC1-DDFD-4A52-84FF-1E978278C760}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
