+++ SpeiderappAPI/Database/createSpeiderappUser.sql
