+++ SpeiderappAPI/Database/init.sh
