+++ SpeiderappAPI/Models/Badge.cs
