+++ SpeiderappAPI/Models/Category.cs
